Add support for array parameters in region files.
Consider a stack A that accepts a `CommaDelimitedList` parameter. Say we have
multiple stacks; B, C, and D; whose outputs we wish to join in a list to
provide to A as this parameter. Previously, there was no way to do this; stack
outputs could only be used as simple parameters. This change allows YAML
arrays to be supplied as parameter values in region files. At resolution time,
the array values are concatenated to form a comma-delimited list. Array
elements that are dictionaries are treated as stack output references, and
resolved.

Signed-off-by: James Hamlin <jfhamlin@cotap.com>